handler

+++ src/main/java/com/indramakers/example/measuresms/services/LocationService.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.indramakers.example.measuresms.config.ErrorCodes;
+import com.indramakers.example.measuresms.exeptions.BusinessException;
 import com.indramakers.example.measuresms.model.entities.Locations;
 import com.indramakers.example.measuresms.model.entities.Measure;
 import com.indramakers.example.measuresms.repositories.LocationRepository;
@@ -27,12 +29,17 @@ public class LocationService {
 
 	public int delete(int id) {
 
-		return locationRepository.delete(id);
+		if (locationRepository.delete(id) == 0) {
+			throw new BusinessException(ErrorCodes.LOCATION_NOT_DELETE);
+		} else {
+
+			return locationRepository.delete(id);
+		}
 
 	}
-	
-	public List<Measure> getMeasureLocation(int idLocation){
-		
+
+	public List<Measure> getMeasureLocation(int idLocation) {
+
 		return locationRepository.getMeasureByLocation(idLocation);
 	}
